Reorganize resourceMapper
