+++ assets/src/containers/folders/ListItemCurrentFolder.tsx
@@ -2,7 +2,8 @@ import { getFolder } from 'actions/folder-actions'
 import LoadingComponent from 'components/Loading'
 import CurrentFolderItem from 'components/lists/CurrentFolderItem'
 import ListItem, { ListItemProps, createDispatchMapper, createResourceMapper } from 'containers/lists/ListItem'
-import createApiResourceWrapper, { ApiResourceDispatch } from 'library/containers/ApiResourceHOC'
+import createApiResourceWrapper from 'library/containers/ApiResourceHOC'
+import { ApiResourceDispatch } from 'library/containers/common'
 import { ApiResourceId } from 'library/service/common'
 import * as React from 'react'
 import { connect } from 'react-redux'
